backend: gameplay route separated, use game session CRUD
