+++ apps/backend/src/router/index.ts
@@ -3,9 +3,11 @@ import { default as gameRoutes } from './game.routes';
 import { default as backendDictionaryRoutes } from './backend/dictionary';
 import { default as backendConfigRoutes } from './backend/config.routes';
 import { default as frontendConfigRoutes } from './frontend/config.routes';
+import { default as frontendGameRoutes } from './frontend/gameplay.routes';
 export {
   gameRoutes,
   backendDictionaryRoutes,
-  frontendConfigRoutes,
   backendConfigRoutes,
+  frontendConfigRoutes,
+  frontendGameRoutes,
 };
